[Misc] Move `GPUModelRunner.prepare_kernel_block_sizes` to utils (#35400)
